Removed the need for a Pivot.
This change shows that there's no need to have a Pivot involved when you can move the Node2D that contains the logic. The cell in which the Actor is to be positioned, is already updated by the Grid. Therefore, it doesn't matter if the position is updated before or after the Tween starts.

This works without issues in Godot 3.2.2.